Add files via upload
Analysis of access to and receipt of medical care during the Covid-19 pandemic. Overall findings include that it appears that care was most accessible in Spring 2021, with care being least accessible in Spring 2020 during the beginning of the pandemic. Additionally, those who were between the ages of 50-59 had the least access to the care they needed.