Add files via upload
This analyzes trends in mental health care during the Covid 19 pandemic. A key finding was that although it appears that Oregon has the lowest access to mental health care and one would be tempted to associate this with a larger portion of residents with Covid 19 cases, some additional research found that the cases were fairly low; in fact, Oregon has a mental health access issue apart from Covid. This is a reminder that findings should be tested against external analyses. Additionally, in general, those with lower education had less access to mental health care. This would likely have been exacerbated by the pandemic due to limited resources and disposable income because of unemployment.